% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxT.R
\name{MinMaxT}
\alias{MinMaxT}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
MinMaxT(data, a, b, alpha)
}
\arguments{
\item{data}{a real data set considering only numerical values}

\item{a}{number of levels of the row factor}

\item{b}{number of levels of the column factor}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
a=3;b=2;N=c(20,30,20,30,40,60);S=c(1,1,2,1,4,2)
g=NULL
for(i in 1:(a*b))
{
 g[[i]]=rnorm(N[i],0,sqrt(S[i]))
}
data=g
MinMaxT(data,3,2,0.05)
}
